nodejs: merge access_token_common into access_token
